requirements.txt
sounddevice>=0.5.0
soundfile>=0.12.0
numpy>=1.24.0
torch>=2.0.0
torchaudio>=2.0.0
nemo_toolkit[asr]
faster-whisper>=1.0.0
textual>=0.89.0
pyyaml>=6.0
